Reorganize integration tests into modular structure
- Split 1688-line integration_tests.rs into organized modules:
  - basic.rs: 10 fundamental endpoint tests
  - http.rs: 4 HTTP method/header/status code tests
  - logging.rs: 2 request logging tests
  - openapi.rs: 8 OpenAPI import/export tests
  - proxy.rs: 8 Smart Proxy Mode tests
  - common.rs: shared TestServer utilities
- Remove all comments from tests for cleaner code
- Total: 32 tests, all properly categorized
- Add proxy source indicators to LogTable UI

This improves test maintainability and organization.
