+++ tests/integration/mod.rs
@@ -1,0 +1,6 @@
+pub mod common;
+pub mod basic;
+pub mod http;
+pub mod logging;
+pub mod openapi;
+pub mod proxy;
